Introduce the run-<architecture>-shell targets.
A run-<architecture>-shell target launches a shell with the QEMU user
mode in a CheriBSD sysroot.

* Since the run-<architecture>-shell targets run a QEMU user mode
binary, different QEMU flags are required, as opposed to the
run-<architecture> targets. Modify the QemuOptions class to provide
separate methods to construct argument lists for a system or a user mode
binary;
* Instead of using the LaunchQEMUBase class, introduce the
LaunchBsdUserQEMUBase class that runs a user mode binary;
* Instead of using the LaunchCheriBSD class, introduce the
LaunchCheriBSDShell class that creates run-<architecture> targets with
the "-shell" suffix;
* Describe the run-<architecture>-shell targets in README.